requirements.txt
rasterio
shapely
numpy
geopandas
pillow
tqdm
mapa
scikit-image
numpy-stl
freetype-py